create package for startscreen and move files
